requirements.txt
# Dependencies
aiogram
python-dotenv


python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1
tenacity==8.2.3
Flask==2.3.2